feat(switch): KISAK_NO_FASTFILES + HunkUser 64-bit pointers + asset seeds
Force the engine into the LoadObj path so we exercise the raw-asset
flow that exists in upstream for the non-fastfile build, get past the
font/menu/material gates with placeholders, and widen HunkUser so the
pointer arithmetic stops truncating to the low 4 GiB on aarch64.

Switch build now defines KISAK_NO_FASTFILES, which makes
IsFastFileLoad() return false everywhere — UI_LoadMenus, asset DB,
sound aliases, etc. all take their `_LoadObj` branches instead of
calling into the (still-unported) .ff parser.

Threading that mode through surfaced a handful of changes:

* universal/com_memory.{h,cpp}: HunkUser::end / HunkUser::pos widened
  from int to uintptr_t. Upstream stores pointers in those slots
  (Hunk_UserCreate does `user->pos = (int)(uintptr_t)user->buf`)
  which truncated to ~28 bits on aarch64 and made Hunk_UserAlloc
  return addresses in the unmapped low 4 GiB. Hunk_UserAlloc's
  alignment math switches to uintptr_t too. The (!(user->pos & 31))
  invariant still holds because we already 32-byte aligned `buf`.

* universal/com_stringtable.cpp:97 cast Hunk_AllocateTempMemoryHigh's
  unsigned-int return through uintptr_t before storing in char* —
  same hex-rays-style 32-bit pointer pattern we keep running into.

* src/switch/switch_renderer_init.cpp:
  - rgp.defaultMaterial pointed at our $default Material so
    Material_MakeDefault doesn't fatal when an unknown name resolves
    to "couldn't load material '$default'".
  - registeredFont[] and registeredFontCount pre-populated with the
    eight font names the engine asks for (fonts/consolefont,
    fonts/consoleFont, fonts/smallfont, fonts/bigfont, ...) so
    R_RegisterFont_LoadObj returns the cached pointer instead of
    walking R_LoadFont's FS_FOpenFileRead path and asserting on the
    missing on-disk file.

* ui/ui_shared_obj.cpp downgrades the two "default.menu file not
  found" Com_Errors to Com_PrintWarning so the engine can keep
  walking past the UI subsystem when no menu file is reachable yet.

Result: Com_Init completes, Com_Frame loop holds steady at 60 FPS
for 2000+ frames with no asserts, no faults. The renderer command
queue still only carries RC_CLEAR_SCREEN / RC_DRAW_PROFILE /
RC_PROJECTION_SET per frame — UI_LoadMenus_LoadObj fails to find
the menu text files via FS_FOpenFileByMode (they live inside IWDs
but the LoadObj path expects loose .menu/.txt on disk), so the
menu list stays empty and there's nothing to RC_STRETCH_PIC.

+++ src/universal/com_memory.h
@@ -61,8 +61,17 @@ struct HunkUser // sizeof=0x24 (32-bit upstream)
     HunkUser* current;
     HunkUser* next;
     int maxSize;
-    int end;
-    int pos;
+    // KISAKHACK-AUDIT(hunkuser-pos-end-64bit): upstream stores `pos` and
+    // `end` as int because in 32-bit they hold full pointers. On aarch64
+    // those pointers don't fit, and the casts in Hunk_UserCreate /
+    // Hunk_UserAlloc lose the high 32 bits, returning addresses in the
+    // unmapped low 4 GiB. Widen both slots to uintptr_t so the engine
+    // sees real addresses; the surrounding hex-rays-style arithmetic
+    // (`alignment & (alignment + pos)`, comparisons, etc.) is still
+    // valid because pointer math is unsigned and the bit patterns
+    // line up.
+    uintptr_t end;
+    uintptr_t pos;
     const char* name;
     bool fixed;
     bool tempMem;
